fix model plane
