+++ src/models/Plane/HandleAnimation.jsx
@@ -1,0 +1,7 @@
+export const HandleAnimation = (actions, isRotating) => {
+  if (isRotating) {
+    actions["Take 001"].play();
+  } else {
+    actions["Take 001"].stop();
+  }
+};
